Remove object

+++ database/sql.py
@@ -1182,8 +1182,7 @@ class removeCircuitObject:
 
         # Insert target object in table of removed objects
         removed_table = facility + '_Removed_CircuitObject'
-        cur.execute( 'INSERT INTO ' + removed_table + ' ( id, room_id, path, zone, voltage_id, object_type, description, parent_id, tail, search_result, source, remove_id ) VALUES(?,?,?,?,?,?,?,?,?,?,?,?) ',
-            ( row[0], row[1], row[2], row[3], row[4], row[5], row[6], row[7], row[8], row[9], row[10], remove_id ) )
+        cur.execute( 'INSERT INTO ' + removed_table + ' ( id, room_id, path, zone, voltage_id, object_type, description, parent_id, tail, search_result, source, remove_id ) VALUES(?,?,?,?,?,?,?,?,?,?,?,?) ', ( *row, remove_id ) )
 
         # Delete target object
         cur.execute( 'DELETE FROM ' + target_table + ' WHERE id=?', ( id, ) )
@@ -1197,7 +1196,7 @@ class removeCircuitObject:
         removed_dev_table = facility + '_Removed_Device'
         for dev in devices:
             device_id = dev[0]
-            cur.execute( 'INSERT INTO ' + removed_dev_table + ' ( id, room_id, parent_id, description, power, name, remove_id ) VALUES(?,?,?,?,?,?,?) ', ( dev[0], dev[1], dev[2], dev[3], dev[4], dev[5], remove_id ) )
+            cur.execute( 'INSERT INTO ' + removed_dev_table + ' ( id, room_id, parent_id, description, power, name, remove_id ) VALUES(?,?,?,?,?,?,?) ', ( *dev, remove_id ) )
             cur.execute( 'DELETE FROM ' + dev_table + ' WHERE id=?', ( device_id, ) )
 
         # Retrieve all descendants of deleted object
@@ -1209,8 +1208,7 @@ class removeCircuitObject:
             descendant_id = desc[0]
 
             # Move current descendant to 'Removed' table
-            cur.execute( 'INSERT INTO ' + removed_table + ' ( id, room_id, path, zone, voltage_id, object_type, description, parent_id, tail, search_result, source, remove_id ) VALUES(?,?,?,?,?,?,?,?,?,?,?,?) ',
-                ( desc[0], desc[1], desc[2], desc[3], desc[4], desc[5], desc[6], desc[7], desc[8], desc[9], desc[10], remove_id ) )
+            cur.execute( 'INSERT INTO ' + removed_table + ' ( id, room_id, path, zone, voltage_id, object_type, description, parent_id, tail, search_result, source, remove_id ) VALUES(?,?,?,?,?,?,?,?,?,?,?,?) ', ( *desc, remove_id ) )
             cur.execute( 'DELETE FROM ' + target_table + ' WHERE id=?', ( descendant_id, ) )
 
             # Retrieve all devices attached to current descendant
@@ -1220,7 +1218,7 @@ class removeCircuitObject:
             # Move all devices attached to current descendant
             for dev in devices:
                 device_id = dev[0]
-                cur.execute( 'INSERT INTO ' + removed_dev_table + ' ( id, room_id, parent_id, description, power, name, remove_id ) VALUES(?,?,?,?,?,?,?) ', ( dev[0], dev[1], dev[2], dev[3], dev[4], dev[5], remove_id ) )
+                cur.execute( 'INSERT INTO ' + removed_dev_table + ' ( id, room_id, parent_id, description, power, name, remove_id ) VALUES(?,?,?,?,?,?,?) ', ( *dev, remove_id ) )
                 cur.execute( 'DELETE FROM ' + dev_table + ' WHERE id=?', ( device_id, ) )
 
         # Log activity
@@ -1259,7 +1257,7 @@ class removeDevice:
 
         # Insert target object in table of removed objects
         removed_table = facility + '_Removed_Device'
-        cur.execute( 'INSERT INTO ' + removed_table + ' ( id, room_id, parent_id, description, power, name, remove_id ) VALUES(?,?,?,?,?,?,?) ', ( row[0], row[1], row[2], row[3], row[4], row[5], remove_id ) )
+        cur.execute( 'INSERT INTO ' + removed_table + ' ( id, room_id, parent_id, description, power, name, remove_id ) VALUES(?,?,?,?,?,?,?) ', ( *row, remove_id ) )
 
         # Delete target object
         cur.execute( 'DELETE FROM ' + target_table + ' WHERE id=?', ( id, ) )
@@ -1297,7 +1295,7 @@ class removeLocation:
 
         # Insert target object in table of removed objects
         removed_table = facility + '_Removed_Room'
-        cur.execute( 'INSERT INTO ' + removed_table + ' ( id, room_num, old_num, location_type, description, remove_id ) VALUES(?,?,?,?,?,?) ', ( row[0], row[1], row[2], row[3], row[4], remove_id ) )
+        cur.execute( 'INSERT INTO ' + removed_table + ' ( id, room_num, old_num, location_type, description, remove_id ) VALUES(?,?,?,?,?,?) ', ( *row, remove_id ) )
 
         # Delete target object
         cur.execute( 'DELETE FROM ' + target_table + ' WHERE id=?', ( id, ) )
@@ -1541,12 +1539,11 @@ class restoreRemovedObject:
         elif remove_object_type == 'Location':
             self.restore_location( by, id, facility )
 
-        # Clean up recyle bin
-        cur.execute( 'DELETE FROM ' + recycle_table + ' WHERE id=?', ( id, ) );
+        if len( self.messages ) == 0:
+            # Clean up recyle bin
+            cur.execute( 'DELETE FROM ' + recycle_table + ' WHERE id=?', ( id, ) );
 
-        conn.commit()
-
-        self.success = True
+            conn.commit()
 
 
     def restore_circuit_object( self, by, id, remove_object_id, parent_id, tail, room_id, facility ):
